R: Rework pdf into webworker
