--- rsconcept/frontend/src/components/pdf/pdf-font-path.ts
@@ -1,4 +1,0 @@
-/** Font URL for @react-pdf in the browser. */
-export function pdfFontSrc(fileName: string): string {
-  return `/fonts/${fileName}`;
-}
