+++ rsconcept/frontend/src/features/rsmodel/pages/rsmodel-page/tab-model-list/tab-model-list.tsx
@@ -113,7 +113,7 @@ export function TabModelList() {
   }
 
   async function createPDFList() {
-    const { cstListToFile } = await import('@/features/rsform/utils/rsform2pdf');
+    const { cstListToFile } = await import('@/services/pdf');
     return cstListToFile(schema.items);
   }
