+++ rsconcept/frontend/src/services/pdf/rsform/index.ts
@@ -1,0 +1,8 @@
+/**
+ * RSForm / constituenta-list PDF export.
+ *
+ * Entry points {@link createSchemaFile} and {@link cstListToFile} are also re-exported from
+ * `@/services/pdf`. Document rendering and the dedicated worker live in this folder; shared PDF
+ * chrome lives one level up under `services/pdf`.
+ */
+export { createSchemaFile, cstListToFile } from './export';
